Readme.md
Design And Implementation of a Cloud-based File sharing system To Create an Online Educational Platform 

This design emphasizes pre-recorded Lecture videos and an attendance system that is taken on the basis of the number of videos the student has watched within the stipulated time. 
The application allows the institutions to dynamically set rules, credits, qualification criteria on the basis of their tradition. 
This design takes off a slot/timetable system allowing students to watch/ rewatch lecture videos at their convenient time. 
It additionally adds a provision that will enable the students to see their daily/ weekly performance and attendance for each week forcing them to plan for the next to come.
It also allows students to connect with other enrolled students and the teaching faculty through a chatbox and a cloud-based video conferencing so that additional doubts about their portions can be clarified.
This system gives a provision of taking proctored assessments and exams to maintain the credibility of the course